+++ packages/runtime/src/parser/parser.ts
@@ -9,7 +9,8 @@ import {
   Btoi, Byte, ByteAdd, ByteBitwiseAnd, ByteBitwiseInvert, ByteBitwiseOr,
   ByteBitwiseXor, Bytec, Bytecblock, ByteDiv, ByteEqualTo, ByteGreaterThanEqualTo, ByteGreatorThan,
   ByteLessThan, ByteLessThanEqualTo, ByteMod, ByteMul, ByteNotEqualTo, ByteSub,
-  ByteZero, Callsub, Concat, Dig, Div, DivModw, Dup, Dup2, Ed25519verify,
+  ByteZero, Callsub, Concat, Dig, Div, DivModw, Dup, Dup2, EcdsaPkDecompress,
+  EcdsaPkRecover, EcdsaVerify, Ed25519verify,
   EqualTo, Err, Exp, Expw, Gaid, Gaids, GetAssetDef, GetAssetHolding,
   GetBit, GetByte, Gload, Gloads, Global, GreaterThan,
   GreaterThanEqualTo, Gtxn, Gtxna, Gtxns, Gtxnsa, Int, Intc, Intcblock, Itob,
@@ -198,6 +199,18 @@ opCodeMap[4] = {
   // Knowable creatable asset
   gaid: Gaid,
   gaids: Gaids
+};
+
+/**
+ * TEALv5 opcodes
+ */
+opCodeMap[5] = {
+  ...opCodeMap[4],
+
+  // ECDSA
+  ecdsa_verify: EcdsaVerify,
+  ecdsa_pk_decompress: EcdsaPkDecompress,
+  ecdsa_pk_recover: EcdsaPkRecover
 };
 
 // list of opcodes that require one extra parameter than others: `interpreter`.